requirements.txt
openai
python-pptx
requests
beautifulsoup4
python-dotenv
